MOBILE-4616 forum: Fix warning when user cannot post to group
We cannot rely only on WS mod_forum_can_add_discussion to decide when
to display the warning because there are multiple reason a user might
not be able to add a discussion.

Now it uses the same logic as in LMS, with exception that the string
"cannotadddiscussiongroup" is used when the user has no groups.
In LMS this string is used when the user does not have the
"canposttomygroups" capability, which doesn't make any sense.